Added studio Module
Commented original create studioo
created studio component
created studio register page
