+++ pages/studio/index.js
@@ -1,15 +1,25 @@
-import Layout from "../../Components/Layout";
-import Image from "next/image";
+import React  from "react";
+import LayoutApp from "../../Components/LayoutApp";
+import FrmStudio from "../../Components/studio/Studiofrm";
 import { useLocalStorage } from "../../hooks/useLocalStorage";
-import Router from 'next/router';
-import {
-    List,
-    ListItem,
-    //Typography,
-    TextField,
-    Button,
-    Alert,
-    AlertTitle,
-    Grid,
-    Snackbar,
-  } from "@mui/material";
+import { useRouter } from "next/router";
+
+const Studio = () => {
+  if (valToken) {
+    const { token, auth, infoUser } = valToken;
+    return (
+      <LayoutApp>
+        <FrmStudio  />
+      </LayoutApp>
+    );
+  } else {
+    return (
+      <LayoutApp>
+        <h1>No autorizado</h1>
+      </LayoutApp>
+    );
+  }
+  
+};
+
+export default Studio
